split and simplify terminal and stop queries
